Reorganized file structure of src directory
